https://www.codechef.com/problems/SNAKPROC
In this problem, the race between snakes and snakes of the kingdom has gathered to participate in the procession. Chef has been tasked with reporting on the procession, and for this, he decides to first keep track of all the snakes. When he sees a snake first, it'll be its Head, and hence he will mark a 'H'. The snakes are long, and when he sees the snake finally slither away, he'll mark a 'T' to denote its tail. In the time in between, when the snake is moving past him, or the time between one snake and the next snake, he marks with '.'s.
Because the snakes come in a procession, and one by one, a valid report would be something like "..H..T…HTH….T.", or "…", or "HT", whereas "T…H..H.T", "H..T..H", "H..H..T..T" would be invalid reports
string=..H..T...HTH....T.  ->valid because it satisfies all conditions like the second snake starts after the first snake ends, which is allowed. here the first snake is first H  T and the second snake is second H SO, (H..H..T..T   -> INVALID)  ".T…H..H.T" is invalid because it has a 'T' before an 'H'. A tail can come only after its head. "H..T..H" is invalid because the last 'H' does not have a corresponding 'T'.